+++ kiari/impl/tool_impl/history_compact/__init__.py
@@ -1,0 +1,3 @@
+from ._models.history_compact import HistoryCompact
+
+__all__ = ["HistoryCompact"]