+++ kiari/impl/tool_impl/text_file_purge/__init__.py
@@ -1,0 +1,3 @@
+from ._models.text_file_purge import TextFilePurge
+
+__all__ = ["TextFilePurge"]